+++ adapters/AdColony/AdColonyAdapter/GADMAdapterAdColonyConstants.h
@@ -2,10 +2,20 @@
 //  Copyright © 2019 Google. All rights reserved.
 //
 
+/// AdColony mediation network adapter version.
 static NSString *const kGADMAdapterAdColonyVersionString = @"3.3.7.0";
 
-static NSString *const kGADMAdapterAdColonyAppIDkey = @"app_id";
+/// AdColony mediation adapter error domain
+static NSString *const kGADMAdapterAdColonyErrorDomain = @"com.google.mediation.adcolony";
 
+/// AdColony mediation adapter zone ID Open bidding key
+static NSString *const kGADMAdapterAdColonyZoneIDOpenBiddingKey = @"zone_id";
+
+/// AdColony mediation adapter zone ID key
 static NSString *const kGADMAdapterAdColonyZoneIDkey = @"zone_ids";
 
-static NSString *const kGADMAdapterAdColonyErrorDomain = @"com.google.mediation.adcolony";
+/// AdColony mediation adapter app ID key
+static NSString *const kGADMAdapterAdColonyAppIDkey = @"app_id";
+
+static NSString *const kGADMAdapterAdColonyAltAdIdKey = @"adc_alt_id";
+static NSString *const kGADMAdapterAdColonyBidReplyKey = @"bid_value";