+++ adapters/AdColony/CHANGELOG.md
@@ -1,5 +1,8 @@
 # AdColony Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Next Version
+- Added open bidding capability to the adapter for interstitial and rewarded ads.
+
 ## Version 3.3.7.0
 - Verified compatibility with AdColony SDK 3.3.7.
 - Fixed a crash in case of failed to fetch rewarded ad.